multi prediction writer
